[Components] flatten components out of folders with styled components.
