--- src/components/MenuWithOptions/styles.scss
@@ -1,6 +1,0 @@
-.menu-opts {
-
-  &-container{
-    max-width: 360px;
-  }
-}
