--- src/components/SettingsFilter/styles.scss
@@ -1,8 +1,0 @@
-.settings-filter {
-  &-line-item{
-    display: flex;
-    flex-direction: row;
-    width: 500px;
-    justify-content: space-between;
-  }
-}
